task: derive GPU-timing pass count from one table, pass GUI panel two spans
GPU_TIMED_PASS_COUNT was hand-written five times; bumping only the count
left the two name tables and pass_ms zero-filled, which reads as a
nullptr pass name in the GUI/JSON export and a bogus 0.000 ms "sample".
GPU_TIMED_PASS_NAMES/EXPORT_NAMES are now std::to_array tables that derive
GPU_TIMED_PASS_COUNT, pinned to a new GpuTimedPass::Count enumerator and to
GpuTimings::pass_ms via static_asserts. renderGpuTimingsPanel takes two
spans instead of a raw pointer + count. Also updated
Compare-RendererTimings.ps1's source regex, which parsed the old C-array
declaration syntax, to accept both spellings.

+++ Test/commit/VulkanEngine/goldenRenderSuite.cpp
@@ -1776,7 +1776,7 @@ TEST(GoldenRender, RaytracingFrameSkipsTheRasterPass)
     const unsigned int forward_meshes_drawn = renderer_vars.visibility.meshes_drawn;
     ASSERT_GT(forward_meshes_drawn, 0U) << "the renderer reported drawing no meshes at all in forward mode";
     const float forward_main_ms =
-      renderer_vars.gpuTimings.pass_ms[static_cast<int>(Kataglyphis::VulkanRendererInternals::FrontendShared::GpuTimedPass::Main)];
+      renderer_vars.gpuTimings.pass_ms[static_cast<size_t>(Kataglyphis::VulkanRendererInternals::FrontendShared::GpuTimedPass::Main)];
 
     renderer_vars.raytracing = true;
     // Flush the rolling window again so the average is dominated by RT-only
@@ -1792,7 +1792,7 @@ TEST(GoldenRender, RaytracingFrameSkipsTheRasterPass)
     if (renderer_vars.gpuTimings.supported) {
         const float rt_main_ms =
           renderer_vars.gpuTimings
-            .pass_ms[static_cast<int>(Kataglyphis::VulkanRendererInternals::FrontendShared::GpuTimedPass::Main)];
+            .pass_ms[static_cast<size_t>(Kataglyphis::VulkanRendererInternals::FrontendShared::GpuTimedPass::Main)];
         GTEST_LOG_(INFO) << "Main-pass GPU time: forward(raster only)=" << forward_main_ms
                           << "ms, RT(dispatch only)=" << rt_main_ms << "ms";
     }
